+++ src/core/requests/tor.py
@@ -84,4 +84,4 @@ def use_tor(request):
   
   return response
   
-#eof
+#eof